Crate silinapse

Silinapse -- A library of silicon synapses

Accross the library, if the input slices do not contain enough values for the unit they are given to, the missing values will be inferred to be 0.0. Samewise, superfluous values are ignored.

The whole library is parametred over a type F, which can be any Float type (currently f32 or f64, but maybe others in the future).

Modules

activations
A set of classic activation functions.

training
A list of available training methods

util
A set of utility method to combine networks.

Structs

BoltzmannMachine
A stochastic self-organizing network.

FeedforwardLayer
A feedforward layer

SymmetricMatrix
Represents a Symmetric Matrix.

Traits

BackpropTrain
A trait for networks that can be trained using a certain method in a back-propagation way: the training returns a values vector that is to be used as a target value for the previous layer.

Compute
A trait representing anything that can process an input to generate an output.

Method
This trait describes a training method. It does not hold a lot of constraints by itself, but networks implementing the same training method should be trainable in the same way.

SupervisedTrain
A trait for networks that can be trained using a certain method of supervised training.

UnsupervisedTrain
A trait for networks that can be trained using a certain method of unsupervised training.