Struct silinapse::FeedforwardLayer

pub struct FeedforwardLayer<F: Float, V: Fn(F) -> F, D: Fn(F) -> F> {
    // some fields omitted
}

A feedforward layer

Such layer is composed of a set of output neurons, and have all its inputs connected to all its outputs.

The effective computation is thus, if X is the vector of inputs, Y the vector of outputs, W the internal weigths matrix, B the vector of biases and f() the activation function (applied on all components of the vector in parallel):

Y = f( W*X + B )

The training of this layer consists on fitting the values of W and B.

Methods

impl<F, V, D> FeedforwardLayer<F, V, D> where F: Float, V: Fn(F) -> F, D: Fn(F) -> F

fn new(inputs: usize, outputs: usize, activation: ActivationFunction<F, V, D>) -> FeedforwardLayer<F, V, D>

Creates a new linear feedforward layer with all its weights set to 0 and its biases set to 0

fn new_from<G>(inputs: usize, outputs: usize, activation: ActivationFunction<F, V, D>, generator: G) -> FeedforwardLayer<F, V, D> where G: FnMut() -> F

Creates a new linear feedforward layer with all its weights and biases generated by provided closure (for example a random number generator).