Struct silinapse::SymmetricMatrix

pub struct SymmetricMatrix<F: Float> {
    // some fields omitted
}

Represents a Symmetric Matrix.

It can be indexed using matrices coordinates like matrix[(i,j)], but for all i and j in range: matrix[(i,j)] and matrix[(j,i)] are the same memory location.

Methods

impl<F: Float> SymmetricMatrix<F>

fn zeros(n: usize) -> SymmetricMatrix<F>

Create a new n x n SymmetricMatrix filled with zeros.

fn size(&self) -> usize

The size of a side of the matrix.

It if returns n, the matrix is thus a n x n matrix.