Struct silinapse::training::GradientDescent

pub struct GradientDescent<F: Float> {
    pub rate: F,
}

The gradient descend approach, consisting on finding a minimum of the error by going down its gradient.

Fields

rate

The learning rate associated with this gradient descent rule.

A very small value will make the training slow, but a too big one will make it unstable.