Trait silinapse::Compute

pub trait Compute<F: Float> {
    fn compute(&self, input: &[F]) -> Vec<F>;
    fn input_size(&self) -> usize;
    fn output_size(&self) -> usize;
}

A trait representing anything that can process an input to generate an output.

This computation is not supposed to alter the internal state of the object.

Required Methods

fn compute(&self, input: &[F]) -> Vec<F>

Process input into output.

fn input_size(&self) -> usize

The number of inputs this network expects.

fn output_size(&self) -> usize

The number of outputs generated by this network.