Trait silinapse::Method

pub trait Method { }

This trait describes a training method. It does not hold a lot of constraints by itself, but networks implementing the same training method should be trainable in the same way.