Trait silinapse::SupervisedTrain

pub trait SupervisedTrain<F: Float, M: Method> {
    fn supervised_train(&mut self, rule: &M, input: &[F], target: &[F]);
}

A trait for networks that can be trained using a certain method of supervised training.

Required Methods

fn supervised_train(&mut self, rule: &M, input: &[F], target: &[F])

Performs one step of supervised training on given input and target value using the learning parameters defined by rule.