Trait silinapse::UnsupervisedTrain

pub trait UnsupervisedTrain<F: Float, M: Method> {
    fn unsupervised_train(&mut self, rule: &M, input: &[F]);
}

A trait for networks that can be trained using a certain method of unsupervised training.

Required Methods

fn unsupervised_train(&mut self, rule: &M, input: &[F])

Performs one step of unsupervised training on given input using the learning parameters defined by rule.