Struct silinapse::util::FixedOutput

pub struct FixedOutput<F: Float> {
    // some fields omitted
}

A network that returns a fixed output, whatever the input is.

Methods

impl<F: Float> FixedOutput<F>

fn new(output: &[F]) -> FixedOutput<F>

Creates a new fixed output network that will always return the values provided as output.