Struct silinapse::util::Parallel

pub struct Parallel<F, A, B> {
    // some fields omitted
}

An adapter that feeds the same input to two networks, and concatenate their outputs into its output.

Methods

impl<F, A, B> Parallel<F, A, B> where F: Float, A: Compute<F>, B: Compute<F>

fn new(first: A, second: B) -> Parallel<F, A, B>

Chains the two given adapters